third_party/blink/web_tests/fast/dom/custom/lifecycle-created-paste-expected.txt

CONSOLE WARNING: line 8: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures that the lifecycle callbacks of editing-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.callbacksCalled is ['X-FOO', 'DIV']
PASS window.callbacksCalled is ['X-FOO', 'DIV', 'X-FOO', 'DIV']
PASS successfullyParsed is true

TEST COMPLETE
